src/main/java/co/com/ceiba/restaurant/dto/ClientDto.java
package co.com.ceiba.restaurant.dto;

public class ClientDto {

	private Integer clientId;
	private String firstName;
	private String lastName;
	private String email;
	private String phoneNumber;
	
	

	public ClientDto() { 
	}

	public ClientDto(Integer clientId, String firstName, String lastName, String email, String phoneNumber) {
		this.clientId = clientId;
		this.firstName = firstName;
		this.lastName = lastName;
		this.email = email;
		this.phoneNumber = phoneNumber;
	}

	public Integer getClientId() {
		return clientId;
	}

	public void setClientId(Integer clientId) {
		this.clientId = clientId;
	}

	public String getFirstName() {
		return firstName;
	}

	public void setFirstName(String firstName) {
		this.firstName = firstName;
	}

	public String getLastName() {
		return lastName;
	}

	public void setLastName(String lastName) {
		this.lastName = lastName;
	}

	public String getEmail() {
		return email;
	}

	public void setEmail(String email) {
		this.email = email;
	}

	public String getPhoneNumber() {
		return phoneNumber;
	}

	public void setPhoneNumber(String phoneNumber) {
		this.phoneNumber = phoneNumber;
	}

}
